third_party/blink/web_tests/fast/frames/take-focus-from-iframe-expected.txt

Focus can be moved around frames.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.activeElement == target is true
PASS target.selectionStart is 0
PASS target.selectionEnd is 0
PASS successfullyParsed is true

TEST COMPLETE
